Feat(Nav) : 유저 데이터 받아와서 이름, 아이콘으로 바꾸기, Footer에서 faq,policy,logo들 a 태그로 href 연결시키기
