--- src/components/aaaaaaaaaaaaa/card.css
@@ -1,44 +1,0 @@
-.card-box {
-    width: 340px;
-    background-color: #fff;
-    border-radius: 15px;
-}
-
-a {
-    color: black;
-}
-
-.card-image{
-    overflow: hidden;
-    width: 340px;
-    height: 253.746px;
-}
-
-.card-text{
-    display: flex;
-    padding: 15px 20px;
-    flex-direction: column;
-    gap: 10px;
-    
-}
-
-.text-description{
-    text-overflow: ellipsis;
-    white-space: normal;
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    color: #000;
-}
-
-.card-wrapper :hover{
-    background-size: 160%;
-}
-
-.card-image {
-    background-size: cover;
-    background-position: center;
-    background-repeat: no-repeat;
-    border-radius: 15px 15px 0px 0px;
-}
